+++ just-plan/src/app/(needLogin)/_components/index.ts
@@ -3,3 +3,4 @@ export * from "./MemoModal/MemoModal";
 export * from "./EditPlanInfoModal/EditPlanInfoModal";
 export * from "./AddPlaceModal/AddPlaceModal";
 export * from "./StoredPlaceMiniCard/StoredPlaceMiniCard";
+export * from "./PlanInfoHeader/PlanInfoHeader";
